python-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
easyocr==1.7.0
pillow==10.1.0
httpx==0.25.0
pydantic==2.5.0
